Reorganise assets
