+++ src/assets/cover_minecraft.png
