+++ src/components/Experience/index.tsx
@@ -1,19 +1,19 @@
 import GridCard from "../GridCard";
 import links from "../../utils/links";
 
-import logo_hexagone from "./assets/ecole_hexagone.svg";
-import logo_luanti from "./assets/luanti.svg";
-import logo_react from "./assets/react.svg";
-import logo_tailwind from "./assets/tailwind.svg";
-import logo_vite from "./assets/vite.svg";
-import logo_supabase from "./assets/supabase.svg";
-import logo_golang from "./assets/golang_aqua.svg";
-import logo_docker from "./assets/docker.svg";
-import logo_kubernetes from "./assets/kubernetes.svg";
-import logo_node from "./assets/node.svg";
-import logo_discord from "./assets/discord.svg";
-import logo_git from "./assets/git.svg";
-import logo_github from "./assets/github.svg";
+import logo_hexagone from "../../assets/ecole_hexagone.svg";
+import logo_luanti from "../../assets/luanti.svg";
+import logo_react from "../../assets/react.svg";
+import logo_tailwind from "../../assets/tailwind.svg";
+import logo_vite from "../../assets/vite.svg";
+import logo_supabase from "../../assets/supabase.svg";
+import logo_golang from "../../assets/golang_aqua.svg";
+import logo_docker from "../../assets/docker.svg";
+import logo_kubernetes from "../../assets/kubernetes.svg";
+import logo_node from "../../assets/node.svg";
+import logo_discord from "../../assets/discord.svg";
+import logo_git from "../../assets/git.svg";
+import logo_github from "../../assets/github.svg";
 
 export default function Experience() {
 	return (
